@@ -2,7 +2,7 @@ import React from "react";
 import { graphql, useFragment } from "react-relay";
 import * as Yup from "yup";
 import { yupResolver } from "@hookform/resolvers/yup";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import useMutationForm from "../../../lib/hooks/useMutationForm";
 import { useSelectCurrentFirm_AccountFragment$key } from "../../../artifacts/useSelectCurrentFirm_AccountFragment.graphql";
 
@@ -57,7 +57,7 @@ export function useSelectCurrentFirm(
       resolver: yupResolver(validationSchema),
       autosave: true,
       onCompleted() {
-        router.reload();
+        router.refresh();
       },
     }
   );
