@@ -1,4 +1,4 @@
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 
 export function useArrayValuesFromQuery(key: string): string[] {
   const router = useRouter();
