@@ -1,5 +1,5 @@
 import React from "react";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import { QueryContext } from "../../components/QueryProvider";
 
 export function useQuery(key?: string, defaultValue?: any) {
