@@ -2,7 +2,7 @@ import { graphql, useFragment } from "react-relay";
 import React from "react";
 import * as Yup from "yup";
 import { yupResolver } from "@hookform/resolvers/yup";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import useMutationForm from "../../../../../lib/hooks/useMutationForm";
 import { useCheckoutForm_BookingFragment$key } from "../../../../../artifacts/useCheckoutForm_BookingFragment.graphql";
 import { useCheckoutForm_CreateCheckoutMutation } from "../../../../../artifacts/useCheckoutForm_CreateCheckoutMutation.graphql";
