@@ -2,7 +2,7 @@ import React from "react";
 import { graphql } from "react-relay";
 import * as Yup from "yup";
 import { yupResolver } from "@hookform/resolvers/yup";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import useMutationForm from "../../../lib/hooks/useMutationForm";
 import { useRemoveFirm_RemoveFirmMutation } from "../../../artifacts/useRemoveFirm_RemoveFirmMutation.graphql";
 
