@@ -5,12 +5,11 @@ import { useTranslation } from "react-i18next";
 import { useCookies } from "react-cookie";
 // @ts-ignore
 import Chatra from "@chatra/chatra";
-import Head from "next/head";
-import Footer from "./Footer";
-import Header from "./Header";
-import { Layout_CurrentUserFragment$key } from "../../artifacts/Layout_CurrentUserFragment.graphql";
-import GlobalSidebarProvider from "../GlobalSidebarProvider";
-import { useApiKey } from "../../lib/hooks/useApiKey";
+import Footer from "components/MainPage/Footer";
+import Header from "components/MainPage/Header";
+import GlobalSidebarProvider from "components/GlobalSidebarProvider";
+import { useApiKey } from "lib/hooks/useApiKey";
+import { Layout_CurrentUserFragment$key } from "artifacts/Layout_CurrentUserFragment.graphql";
 
 type LayoutProps = {
   children:
@@ -20,27 +19,17 @@ type LayoutProps = {
     | React.ReactNode[];
   currentUserFragment: Layout_CurrentUserFragment$key;
   showRegisterFirm?: boolean;
-  CSN?: boolean;
-  title?: string;
 };
 
 const Layout = ({
   children,
   currentUserFragment,
   showRegisterFirm = true,
-  CSN = false,
-  title,
 }: LayoutProps) => {
   const [isSSR, setIsSSR] = React.useState(true);
 
   React.useEffect(() => {
     setIsSSR(false);
-  }, []);
-
-  React.useEffect(() => {
-    if (CSN) {
-      window.location.reload();
-    }
   }, []);
 
   const currentUser = useFragment(
@@ -56,7 +45,7 @@ const Layout = ({
     `,
     currentUserFragment
   );
-  const { i18n, t } = useTranslation();
+  const { i18n } = useTranslation();
   const [value] = useCookies();
 
   React.useEffect(() => {
@@ -73,34 +62,17 @@ const Layout = ({
     Chatra("init", config);
   }, [chatraApiKey]);
 
-  const titleString = React.useMemo(() => {
-    if (title) return title;
-    if (currentUser?.account?.firm) {
-      return t("layout.metadata.firmTitle");
-    }
-    return t("layout.metadata.commonTitle");
-  }, [currentUser]);
-
-  if (CSN) {
-    return null;
-  }
-
   return (
-    <>
-      <Head>
-        <title>{titleString}</title>
-      </Head>
-      <GlobalSidebarProvider firmFragmentRef={currentUser?.account?.firm!}>
-        <Sheet>
-          <Header
-            currentUserFragment={currentUser}
-            showRegisterFirm={showRegisterFirm}
-          />
-          {!isSSR ? children : null}
-          <Footer />
-        </Sheet>
-      </GlobalSidebarProvider>
-    </>
+    <GlobalSidebarProvider firmFragmentRef={currentUser?.account?.firm!}>
+      <Sheet>
+        <Header
+          currentUserFragment={currentUser}
+          showRegisterFirm={showRegisterFirm}
+        />
+        {!isSSR ? children : null}
+        <Footer />
+      </Sheet>
+    </GlobalSidebarProvider>
   );
 };
 
